fixed the code to elminate warnings by clang (one was clearly a bug).

+++ src/ras/distrasci.cc
@@ -130,7 +130,7 @@ void DistRASCI::generate_guess(const int nspin, const int nstate, shared_ptr<Dis
     const double fac = adapt.second;
     for (auto& iter : adapt.first) {
       shared_ptr<DistRASBlock<double>> block = out->data(oindex)->block(get<0>(iter), get<1>(iter));
-      const size_t aindex = block->stringa()->lexical<0>(get<1>(iter)) - block->astart();
+      const int aindex = block->stringa()->lexical<0>(get<1>(iter)) - block->astart();
       if ( aindex >= 0 && aindex < block->asize()) {
         const size_t bindex = block->stringb()->lexical<0>(get<0>(iter));
         double* data = block->local() + block->lenb() * aindex + bindex;
